+++ src/apis/property/review/fetchReviewList.ts
@@ -1,36 +1,5 @@
 import axiosInstance from "@/apis/utils/axiosInstance";
-
-export type HasChildrenStatus = "NON_CHILDREN" | "HAS_CHILDREN";
-export type ResidentStatus = "NON_RESIDENT" | "CURRENT_RESIDENT" | "PAST_RESIDENT";
-
-export type Review = {
-  reviewId: number;
-  nickname: string;
-  profileImage: string | null;
-  rating: number;
-  content: string;
-  hasChildren: HasChildrenStatus;
-  isResident: ResidentStatus;
-  createdAt: string;
-  updatedAt: string | null;
-  likeCount: number;
-  commentCount: number;
-  isLikedByMe: boolean;
-  isMine: boolean;
-};
-
-export type ReviewListData = {
-  propertyId: number;
-  complexId: number | null;
-  avgRating: number;
-  reviews: Review[];
-};
-
-export type ReviewListResponse = {
-  status: number;
-  message: string;
-  data: ReviewListData;
-};
+import { ReviewListData, ReviewListResponse } from "../../../types/reviewType";
 
 export const fetchReviewList = async (
   propertyId: number,
