+++ src/apis/property/review/fetchReviewSummary.ts
@@ -1,20 +1,5 @@
 import axiosInstance from "@/apis/utils/axiosInstance";
-
-export type ReviewSummary = {
-  good: string[];
-  bad: string[];
-  tra: string[];
-  edu: string[];
-  hel: string[];
-  loc: string[];
-};
-
-export type ReviewSummaryResponse = {
-  status: number;
-  result: boolean;
-  message: string;
-  data: ReviewSummary;
-};
+import { ReviewSummary, ReviewSummaryResponse } from "../../../types/reviewType";
 
 export const fetchReviewSummary = async (propertyId: number): Promise<ReviewSummary> => {
   const response = await axiosInstance.get<ReviewSummaryResponse>(`/reviews/${propertyId}/summary`);
