+++ src/apis/property/review/putReviewLike.ts
@@ -1,15 +1,10 @@
 import axiosInstance from "@/apis/utils/axiosInstance";
-
-export type ToggleReviewLikeResponse = {
-  reviewId: number;
-  userId: number;
-  isLiked: boolean;
-};
+import { ToggleReviewLikeType } from "@/types/reviewType";
 
 export const toggleReviewLike = async (
   reviewId: number,
   isLiked: boolean,
-): Promise<ToggleReviewLikeResponse> => {
+): Promise<ToggleReviewLikeType> => {
   const response = await axiosInstance.put(`/reviews/${reviewId}/likes`, { isLiked });
   return response.data.data;
 };
